terminal/requirements.txt
RPi.GPIO==0.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyserial==3.5
requests==2.25.1
urllib3==1.26.2
